+++ modules/ffmpeg/producer/util/util.h
@@ -69,7 +69,7 @@ AVRational fix_time_base(AVRational time_base);
 
 double read_fps(AVFormatContext& context, double fail_value);
 
-std::wstring print_mode(size_t width, size_t height, double fps, bool interlaced);
+std::wstring print_mode(int width, int height, double fps, bool interlaced);
 
 std::wstring probe_stem(const std::wstring stem);
 bool is_valid_file(const std::wstring filename);
